IOSS: Separate chain generation function into IOSS for general use
